+++ internal/version/init.go
@@ -1,3 +1,3 @@
 package version
 
-const VERSION = "v0.6.3"
+const VERSION = "v0.7.0-beta"